+++ backend/backend/src/main/java/backend/controller/UserController.java
@@ -141,15 +141,7 @@ public class UserController {
 	
 	
 	
-	 //update slike usera razbijanje requesta na multipart i json deo
-	@PreAuthorize("hasAnyRole('ROLE_ADMIN','ROLE_REGISTERED_USER')")
-	@PutMapping(value = "/image",produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-	public ResponseEntity<User> updateImage(@RequestParam("file")MultipartFile file, Principal principal) {
-		User user = userService.findByUsername(principal.getName());
-		user.setImageUrl(fileUploadService.imageUpload(file));
-		userService.save(user);
-		return new ResponseEntity<>(user, HttpStatus.OK);
-	}
+
